pyproject.toml
[project]
name = "cov-steer-stl"
version = "0.1.0"
description = "Differentiable Covariance Steering with Probabilistic Signal Temporal Logic"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "scipy>=1.10",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]

[tool.ruff]
line-length = 100
